+++ swagger_parser/lib/src/generator/fill_controller.dart
@@ -16,6 +16,7 @@ final class FillController {
     bool putClientsInFolder = false,
     bool freezed = false,
     bool enumsToJson = false,
+    bool unknownEnumValue = true,
     bool markFilesAsGenerated = false,
   })  : _openApiInfo = openApiInfo,
         _clientPostfix = clientPostfix,
@@ -24,6 +25,7 @@ final class FillController {
         _putClientsInFolder = putClientsInFolder,
         _freezed = freezed,
         _enumsToJson = enumsToJson,
+        _unknownEnumValue = unknownEnumValue,
         _markFilesAsGenerated = markFilesAsGenerated;
 
   final OpenApiInfo _openApiInfo;
@@ -33,6 +35,7 @@ final class FillController {
   final bool _freezed;
   final bool _putClientsInFolder;
   final bool _enumsToJson;
+  final bool _unknownEnumValue;
   final bool _markFilesAsGenerated;
 
   /// Return [GeneratedFile] generated from given [UniversalDataClass]
@@ -44,6 +47,7 @@ final class FillController {
           dataClass,
           freezed: _freezed,
           enumsToJson: _enumsToJson,
+          unknownEnumValue: _unknownEnumValue,
           markFilesAsGenerated: _markFilesAsGenerated,
         ),
       );